+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ensure cluster pool destroy() gracefully shutdowns worker's server.
 - Ensure pool event is emitted before task error promise rejection.
+- Fix queued tasks count computation.
 
 ## [2.6.3] - 2023-06-19